Add Watch method to agent cache that abstracts local blocking on the cache.
It gets really hard to generalise this outside of the cache package due to request and response types being specific to each case.

It's quite gross that we now have 3 different blocking loops at play in the cache package but having tried a few times to writing external components that block on local cache values and keep hitting gross parts it felt better to have this in the cache package once with a better interface for other components to consume without worrying about backoffs and blocking details.

I feel like it might be worth refactoring the cache at some point to simplify this but not now...